Camera & math updates.
Moved the pi constants to the UtilMath header and added a few helper functions and constants.
Updated both graphics classes to not update the graphics presentation when the height property is zero.
Refactored and removed unnecessary logic.
Changed the pitch polarity.
Changed how the camera's calculate the lookat vector so that it is done on demand rather than cached.
Updated the DirectX camera to rotate counter-clockwise to be more inline with the RHS system.
Updated the OpenGL camera to support the roll-pitch-yaw logic similiar to the DirectX camera.
Updated the windows resource file to set DirectX as checked in the menu.